+++ tests/testthat/test_bnpsd.R
@@ -1458,6 +1458,102 @@ test_that("draw_all_admix with provided p_anc (vector) works", {
     expect_equal( out$p_anc, p_anc )
 })
 
+test_that("draw_all_admix works with p_anc_distr", {
+    m_loci <- 10
+    inbr_subpops <- c(0.1, 0.2, 0.3)
+    k_subpops <- length(inbr_subpops)
+    names( inbr_subpops ) <- paste0( 'S', 1 : k_subpops )
+    admix_proportions <- diag(rep.int(1, k_subpops)) # island model for test...
+    # repeat so we have multiple people per island
+    admix_proportions <- rbind(admix_proportions, admix_proportions, admix_proportions)
+    n_ind <- nrow(admix_proportions) # number of individuals (3*k_subpops)
+    colnames( admix_proportions ) <- names( inbr_subpops )
+    rownames( admix_proportions ) <- paste0( 'i', 1 : n_ind )
+    # in all cases use a very high maf_min to force re-draws, to make sure that works too
+    maf_min <- 8/18
+    # construct p_anc_distr
+    # since it's just a distribution, it can have any length (shouldn't cause errors)
+    # (chose 17, which has no common factors with 10)
+    p_anc_distr <- runif( 17 )
+    
+    # run draw_all_admix, only test default (p_ind and p_subpops not returned)
+    out <- draw_all_admix(admix_proportions, inbr_subpops, m_loci, maf_min = maf_min, p_anc_distr = p_anc_distr)
+    expect_equal( names(out), draw_all_admix_names_ret_default )
+    X <- out$X # genotypes
+    p_anc <- out$p_anc # Ancestral AFs
+    # test X
+    expect_equal(nrow(X), m_loci)
+    expect_equal(ncol(X), n_ind)
+    expect_true( !anyNA( X ) ) # no missing values
+    expect_true(all(X %in% c(0, 1, 2))) # only three values allowed!
+    expect_true(!any(fixed_loci(X))) # we don't expect any loci to be fixed
+    # only individuals have names!
+    expect_equal( colnames( X ), rownames( admix_proportions ) )
+    # test p_anc
+    expect_equal(length(p_anc), m_loci)
+    expect_true(all(p_anc >= 0)) # all are non-negative
+    expect_true(all(p_anc <= 1)) # all are smaller or equal than 1
+    # p_anc was simulated inside function (not passed) so loci don't have names
+    expect_true( is.null( names( p_anc ) ) )
+
+    # repeat with a distribution with fewer elements than m_loci
+    p_anc_distr <- runif( 3 )
+    out <- draw_all_admix(admix_proportions, inbr_subpops, m_loci, maf_min = maf_min, p_anc_distr = p_anc_distr)
+    expect_equal( names(out), draw_all_admix_names_ret_default )
+    X <- out$X # genotypes
+    p_anc <- out$p_anc # Ancestral AFs
+    # test X
+    expect_equal(nrow(X), m_loci)
+    expect_equal(ncol(X), n_ind)
+    expect_true( !anyNA( X ) ) # no missing values
+    expect_true(all(X %in% c(0, 1, 2))) # only three values allowed!
+    expect_true(!any(fixed_loci(X))) # we don't expect any loci to be fixed
+    expect_equal( colnames( X ), rownames( admix_proportions ) )
+    # test p_anc
+    expect_equal(length(p_anc), m_loci)
+    expect_true(all(p_anc >= 0)) # all are non-negative
+    expect_true(all(p_anc <= 1)) # all are smaller or equal than 1
+    expect_true( is.null( names( p_anc ) ) )
+
+    # edge case: does scalar work?
+    p_anc_distr <- 0.3
+    out <- draw_all_admix(admix_proportions, inbr_subpops, m_loci, maf_min = maf_min, p_anc_distr = p_anc_distr)
+    expect_equal( names(out), draw_all_admix_names_ret_default )
+    X <- out$X # genotypes
+    p_anc <- out$p_anc # Ancestral AFs
+    # test X
+    expect_equal(nrow(X), m_loci)
+    expect_equal(ncol(X), n_ind)
+    expect_true( !anyNA( X ) ) # no missing values
+    expect_true(all(X %in% c(0, 1, 2))) # only three values allowed!
+    expect_true(!any(fixed_loci(X))) # we don't expect any loci to be fixed
+    expect_equal( colnames( X ), rownames( admix_proportions ) )
+    # test p_anc
+    expect_equal(length(p_anc), m_loci)
+    expect_true(all(p_anc >= 0)) # all are non-negative
+    expect_true(all(p_anc <= 1)) # all are smaller or equal than 1
+    expect_true( is.null( names( p_anc ) ) )
+
+    # repeat with a function!
+    p_anc_distr <- runif
+    out <- draw_all_admix(admix_proportions, inbr_subpops, m_loci, maf_min = maf_min, p_anc_distr = p_anc_distr)
+    expect_equal( names(out), draw_all_admix_names_ret_default )
+    X <- out$X # genotypes
+    p_anc <- out$p_anc # Ancestral AFs
+    # test X
+    expect_equal(nrow(X), m_loci)
+    expect_equal(ncol(X), n_ind)
+    expect_true( !anyNA( X ) ) # no missing values
+    expect_true(all(X %in% c(0, 1, 2))) # only three values allowed!
+    expect_true(!any(fixed_loci(X))) # we don't expect any loci to be fixed
+    expect_equal( colnames( X ), rownames( admix_proportions ) )
+    # test p_anc
+    expect_equal(length(p_anc), m_loci)
+    expect_true(all(p_anc >= 0)) # all are non-negative
+    expect_true(all(p_anc <= 1)) # all are smaller or equal than 1
+    expect_true( is.null( names( p_anc ) ) )
+})
+
 test_that( "validate_coanc_tree works", {
     # draw a random tree that is valid
     k_subpops <- 5